- fandogh_presenter package created. - table_presenter pyhton file created. - generate_service_detail method moved to table_presenter
